Rework module generation
Instead of relying on the `Code` module, generate a form and run that
through Erlang's `compile:forms/2` function. With that, we can ensure
that line numbers can match the markdown file checked, we do not need to
disable Elixir compiler warnings, and this gives us the nice feeling of
having written some Erlang as well.